parse number as floats, cast to int

+++ src/query.rs
@@ -22,7 +22,7 @@ pub fn query(combined: PathBuf, region: String) {
                 let values = splitline
                     .iter()
                     .skip(3)
-                    .map(|number| number.parse::<i32>().expect("Failed parsing lengths"));
+                    .map(|number| number.parse::<f64>().expect("Failed parsing lengths") as i64);
                 let mut zipped = samples.zip(values).into_iter().collect::<Vec<_>>();
                 zipped.sort_by_key(|&(_, val)| -val);
                 for (name, val) in zipped {
